Submitting Assignment 9.15
First Processing Game: CanniBALLS
This game is inspired by big fish eating the little fish in the sea. Why can't the big balls eat the little balls of the computer screen. You are the little ball and must roll away. Keep in mind, the bigger balls have a wider circumference than you - they can roll faster... This is where the strategy comes in.