Submitting Assignment 11.15
This is a p5.js app which reads in data from my previous project DAQS. DAQS is an air quality monitoring device which remotely collects data from an array of sensors and uploads it in real time to Firebase. This is a current work in progress. I hope to develop it so it is formatted in a way that the information can be easily interpreted and manipulated and develop features so the user can pan in time and also select which device and which parameter to show.