Submitting Final Project
This visualization is based on the data of another project I am currently working on independent of classes:
The DAQS system uses the existing bus network to provide mobile sensing over a repetitive and suitable route. Two types of air quality units are used: high-sensitivity units fitted with high-grade electrochemical sensors and low-cost units fitted with cheaper electrochemical sensors that have been used successfully in other studies for air quality monitoring. On selected routes, buses are fitted with high-sensitivity units and the bus stops on those routes are fitted with low-cost units. The bus stops provide the needed coverage and real time data for the system while the buses serve as a calibration reference each time they pass a bus-stop. This is because data from both sets of sensors are uploaded to a database in real time and the data is calibrated then processed before making the information available to end users. This achieves accuracies comparable to the more expensi